clean up worker interface
