+++ main/signers/hot/SeedSigner/worker.ts
@@ -12,7 +12,7 @@ import {
 import type {
   CoreWorkerMethod,
   EncryptSeedParams,
-  HotSignerWorker,
+  HotSignerMessageHandler,
   PseudoCallback,
   SeedSignerMethod,
   SignMessageParams,
@@ -32,7 +32,7 @@ function isSeedSignerMethod(method: string): method is CoreWorkerMethod | SeedSi
   return isHotSignerMethod(method) || ['encryptSeed'].includes(method)
 }
 
-export default class SeedSignerWorker implements HotSignerWorker {
+export default class SeedSignerWorker implements HotSignerMessageHandler {
   private seed: string | null = null
 
   handleMessage(cb: PseudoCallback<unknown>, method: WorkerMethod, params: any) {
